requirements.txt
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
openai>=1.3.0
pypdf>=3.17.0
python-docx>=1.1.0
tiktoken>=0.5.1
numpy>=1.24.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0

